Update GameService with more checks

+++ src/main/java/ch/fhnw/wodss/tippspiel/Services/GameService.java
@@ -1,7 +1,9 @@
 package ch.fhnw.wodss.tippspiel.Services;
 
 import ch.fhnw.wodss.tippspiel.Domain.Game;
+import ch.fhnw.wodss.tippspiel.Exception.IllegalActionException;
 import ch.fhnw.wodss.tippspiel.Exception.ResourceNotFoundException;
+import ch.fhnw.wodss.tippspiel.Persistance.BetRepository;
 import ch.fhnw.wodss.tippspiel.Persistance.GameRepository;
 import ch.fhnw.wodss.tippspiel.Persistance.TournamentGroupRepository;
 import ch.fhnw.wodss.tippspiel.Persistance.TournamentTeamRepository;
@@ -9,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.Calendar;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,13 +21,15 @@ public class GameService {
     private final GameRepository gameRepository;
     private final TournamentTeamRepository tournamentTeamRepository;
     private final TournamentGroupRepository tournamentGroupRepository;
+    private final BetRepository betRepository;
 
 
     @Autowired
-    public GameService(GameRepository gameRepository, TournamentTeamRepository tournamentTeamRepository, TournamentGroupRepository tournamentGroupRepository) {
+    public GameService(GameRepository gameRepository, TournamentTeamRepository tournamentTeamRepository, TournamentGroupRepository tournamentGroupRepository, BetRepository betRepository) {
         this.gameRepository = gameRepository;
         this.tournamentTeamRepository = tournamentTeamRepository;
         this.tournamentGroupRepository = tournamentGroupRepository;
+        this.betRepository = betRepository;
     }
 
     public List<Game> getAllGames() {
@@ -37,6 +42,9 @@ public class GameService {
     }
 
     public Game addGame(Game game) {
+        if (gameRepository.existsGameByHomeTeamAndAwayTeamAndDateTimeEquals(game.getHomeTeam(), game.getAwayTeam(), game.getDateTime())) {
+            throw new IllegalActionException("Can't create an identical game");
+        }
         return gameRepository.save(game);
     }
 
@@ -50,6 +58,9 @@ public class GameService {
     }
 
     public void deleteGame(Long id) {
+        if (betRepository.existsBetsByGame_Id(id)) {
+            throw new IllegalActionException("Can't delete a game with active bets");
+        }
         if (gameRepository.existsById(id)) {
             gameRepository.deleteById(id);
         } else {
@@ -59,7 +70,11 @@ public class GameService {
 
     public void setResult(Long id, int homeTeamScore, int awayTeamScore) {
         Optional<Game> game = gameRepository.findById(id);
+        Calendar cal = Calendar.getInstance();
         if (game.isPresent()) {
+            if (game.get().getDateTime().before(cal.getTime())) {
+                throw new IllegalActionException("Can't set Score for a game which hasn't been played jet.");
+            }
             game.get().setHomeTeamGoals(homeTeamScore);
             game.get().setAwayTeamGoals(awayTeamScore);
             gameRepository.save(game.get());
